+++ Juno/Week5/Week5_Mission/src/main/java/com/example/week5_mission/domain/review/dto/ReviewReqDTO.java
@@ -1,11 +1,4 @@
 package com.example.week5_mission.domain.review.dto;
 
-import lombok.Getter;
-
 public class ReviewReqDTO {
-
-    @Getter
-    public static class PostReview {
-        private String content;
-    }
 }
