+++ Juno/Week5/Week5_Mission/src/main/java/com/example/week5_mission/domain/review/dto/ReviewResDTO.java
@@ -1,7 +1,4 @@
 package com.example.week5_mission.domain.review.dto;
 
 public class ReviewResDTO {
-    public static class Review {
-        Long reviewId;
-    }
 }
