+++ Juno/Week5/Week5_Mission/src/main/java/com/example/week5_mission/domain/store/dto/StoreReqDTO.java
@@ -1,4 +1,12 @@
 package com.example.week5_mission.domain.store.dto;
 
+import lombok.Getter;
+
 public class StoreReqDTO {
+    
+
+    @Getter
+    public static class PostReview {
+        private String content;
+    }
 }
